Add contours detection + canvas

+++ HideoutBoard/android/app/src/main/java/com/reactlibrary/RNOpenCvLibraryModule.java
@@ -1,42 +1,31 @@
 package com.reactlibrary;
 
+import android.content.res.AssetManager;
+import android.os.Build;
+import android.util.Log;
+
+import androidx.annotation.RequiresApi;
+
+import com.facebook.react.bridge.Callback;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
-import com.facebook.react.bridge.Callback;
 import com.google.gson.Gson;
 
-
-import android.content.Context;
-import android.content.res.AssetManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-
 import org.opencv.core.Core;
-import org.opencv.core.CvType;
 import org.opencv.core.Mat;
-
-import org.opencv.android.Utils;
-import org.opencv.core.MatOfByte;
 import org.opencv.core.MatOfFloat;
 import org.opencv.core.MatOfInt;
 import org.opencv.core.MatOfPoint;
 import org.opencv.core.MatOfRect2d;
 import org.opencv.core.Point;
-import org.opencv.core.Rect;
 import org.opencv.core.Rect2d;
 import org.opencv.core.Scalar;
 import org.opencv.core.Size;
-import org.opencv.dnn.Net;
-import org.opencv.imgproc.Imgproc;
-import org.opencv.imgcodecs.Imgcodecs;
 import org.opencv.dnn.Dnn;
-
-import android.os.Build;
-import android.util.Base64;
-import android.util.Log;
-
-import androidx.annotation.RequiresApi;
+import org.opencv.dnn.Net;
+import org.opencv.imgcodecs.Imgcodecs;
+import org.opencv.imgproc.Imgproc;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -45,6 +34,8 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
 
@@ -102,7 +93,8 @@ public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
 
             net.forward(outs, output_layers);
 
-            System.out.println("[DEBUG] : Detecting objects Done");
+            net = null;
+            blob = null;
 
             List<Rect2d>  boxes= new ArrayList<Rect2d>();
             List<Float>  confidences= new ArrayList<Float>();
@@ -139,39 +131,65 @@ public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
 
             Dnn.NMSBoxes(matBoxes, matConfidences, (float)0.5, (float)0.4, mnsIndexes);
 
+            confidences = null;
+            matConfidences = null;
+
             List<Double[]> holds= new ArrayList<Double[]>();
             List<List<Point>> holdsContours = new ArrayList<List<Point>>();
-//            for(int i: mnsIndexes.toList()) {
-                int i =  mnsIndexes.toList().get(3);
-                // Contours detection
-                int x = (int)boxes.get(i).x;
-                int y = (int)boxes.get(i).y;
-                int w = (int)boxes.get(i).width;
-                int h = (int)boxes.get(i).height;
-                Mat frame = resizedImgMat.submat(y, y+h, x, x+w);
+            int step = 100;
+            double maxArea = 0.0;
+            int x;
+            int y;
+            int w;
+            int h;
+
+
+
+            for(int i: mnsIndexes.toList()) {
+//                int i =  mnsIndexes.toList().get(3);
+                System.gc();
+                System.runFinalization();
+
+                x = (int)boxes.get(i).x;
+                y = (int)boxes.get(i).y;
+                w = (int)boxes.get(i).width;
+                h = (int)boxes.get(i).height;
+
+                Mat frame;
+                try{
+                    frame = resizedImgMat.submat(y, y + h, x, x + w);
+                }
+                catch (Exception e) {
+                    continue;
+                }
 
                 // TODO: Add brightness contrast
+                frame = applyBrightnessContrast(frame, 64, 64);
+                System.gc();
+                System.runFinalization();
 
                 Imgproc.GaussianBlur(frame, frame, new Size(5,5), 0);
                 Imgproc.cvtColor(frame, frame, Imgproc.COLOR_BGR2HSV);
 
                 Mat color_frame = frame.submat(h/4, h - (h/4), w/4, w - (w/4));
                 Scalar avgColor = Core.mean(color_frame);
+                color_frame.release();
 
-                Mat mask = frame;
-                int step = 100;
+                Mat mask = new Mat();
                 Scalar lower = new Scalar(avgColor.val[0] - step, avgColor.val[1] - step, avgColor.val[2] - step, avgColor.val[3]);
                 Scalar upper = new Scalar(avgColor.val[0] + step, avgColor.val[1] + step, avgColor.val[2] + step, avgColor.val[3]);
                 Core.inRange(frame, lower, upper, mask);
+                frame.release();
 
                 List<MatOfPoint> contours = new ArrayList<>();
                 Mat hierarchy = new Mat();
                 Imgproc.findContours(mask, contours, hierarchy, Imgproc.RETR_TREE, Imgproc.CHAIN_APPROX_NONE);
+                mask.release();
 
-                double maxArea = 0.0;
                 MatOfPoint maxContour = new MatOfPoint();
-                maxContour.fromArray(new Point(x,y), new Point(x+w,y), new Point(x,y+h), new Point(x+w,y+h));
+                maxContour.fromArray(new Point(x,y), new Point(x+w,y), new Point(x+w,y+h), new Point(x,y+h));
 
+                maxArea = 0.0;
                 for (MatOfPoint contour: contours) {
                     double area = Imgproc.contourArea(contour);
                     if  (area > maxArea && area <= ( w*h*0.98 ) && area >= (w*h*0.1)) {
@@ -180,20 +198,43 @@ public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
                     }
                 }
 
+//                List<Point> maxContourToList = maxContour.toList();
+//                maxContour = null;
+//                smooth_step = maxContourToList.size() / 16;
+//                List<Point> smooth_contour = new ArrayList<Point>();
+//                for (int batch = 0; batch < maxContourToList.size(); batch+=smooth_step ) {
+//                    Point point = maxContourToList.get(batch);
+//                    point.x += x;
+//                    point.y += y;
+//                    smooth_contour.add(point);
+//                }
+//                maxContourToList = null;
+
                 List<Point> maxContourToList = maxContour.toList();
-                int smooth_step = maxContourToList.size() / 10;
-                List<Point> smooth_contour = new ArrayList<Point>();
-                for (int batch = 0; batch < maxContourToList.size(); batch+=smooth_step ) {
-                    Point point = maxContourToList.get(batch);
-                    point.x += x;
-                    point.y += y;
-                    smooth_contour.add(point);
+                maxContour  = null;
+                int skip = 12;
+                List<Point> smooth_contour = maxContourToList;
+                if (maxContourToList.size() >= skip) {
+                    int smooth_step = maxContourToList.size() / skip;
+                    int limit = maxContourToList.size() / smooth_step + Math.min(maxContourToList.size() % smooth_step, 1);
+                    smooth_contour = Stream.iterate(0, j -> j + smooth_step)
+                            .limit(limit)
+                            .map(maxContourToList::get)
+                            .collect(Collectors.toList());
+                    maxContourToList = null;
                 }
 
+                for( Point p: smooth_contour){
+                    p.x += x;
+                    p.y += y;
+                }
+
+
                 holdsContours.add(smooth_contour);
+                smooth_contour = null;
                 holds.add(new Double[] {boxes.get(i).x, boxes.get(i).y, boxes.get(i).width, boxes.get(i).height});
 
-//            }
+            }
 
 
             successCallback.invoke(new Gson().toJson(holds), new Gson().toJson(holdsContours));
@@ -221,6 +262,10 @@ public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
             try {
                 in = assetManager.open("yolo/" + filename);
                 File outFile = new File(this.reactContext.getDataDir() +"/", filename);
+                if(outFile.exists()) {
+                    System.out.println("[DEBUG] file " + this.reactContext.getDataDir() +"/" + filename + "already existe");
+                    continue;
+                }
                 out = new FileOutputStream(outFile);
                 copyFile(in, out);
                 in.close();
@@ -236,6 +281,34 @@ public class RNOpenCvLibraryModule extends ReactContextBaseJavaModule {
         System.out.println("[DEBUG] Finished coping assets to chache");
     }
 
+    private Mat applyBrightnessContrast(Mat inputImg, int brightness, int contrast) {
+        Mat buf = new Mat();
+        if (contrast != 0) {
+            int highlight = 255;
+            int shadow = brightness;
+            if (brightness < 0) {
+                highlight += brightness;
+                shadow = 0;
+            }
+            float alpha_b = (highlight - shadow) / 255;
+            float gamma_b = shadow;
+            Core.addWeighted(inputImg, alpha_b, inputImg, 0, gamma_b, buf);
+        }
+        else {
+            buf= inputImg;
+        }
+
+        if (contrast != 0) {
+            float f = 131*(contrast + 127)/(127*(131-contrast));
+            float alpha_c = f;
+            float gamma_c = 127*(1-f);
+
+            Core.addWeighted(buf,  alpha_c, buf, 0, gamma_c, buf);
+        }
+
+        return inputImg;
+
+    }
 
     private void copyFile(InputStream in, OutputStream out) throws IOException {
         byte[] buffer = new byte[1024];
